# infra/openapi.yaml
openapi: 3.0.1
info:
  title: GitHub Issues Summary
  version: v1
servers:
  - url: https://localhost:5001
paths:
  /api/Chat/completions:
    post:
      tags:
        - Chat
      operationId: ChatCompletions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/GitHub/issues:
    get:
      tags:
        - GitHub
      operationId: Issues
      parameters:
        - name: User
          in: query
          required: true
          schema:
            type: string
        - name: Repository
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GitHubIssueCollectionResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/GitHubIssueCollectionResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/GitHubIssueCollectionResponse'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/GitHub/issues/{id}:
    get:
      tags:
        - GitHub
      operationId: IssueById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: User
          in: query
          required: true
          schema:
            type: string
        - name: Repository
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GitHubIssueItemResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/GitHubIssueItemResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/GitHubIssueItemResponse'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/GitHub/issues/{id}/summary:
    get:
      tags:
        - GitHub
      operationId: IssueSummaryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: User
          in: query
          required: true
          schema:
            type: string
        - name: Repository
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GitHubIssueItemSummaryResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/GitHubIssueItemSummaryResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/GitHubIssueItemSummaryResponse'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ChatCompletionRequest:
      required:
        - prompt
      type: object
      properties:
        prompt:
          minLength: 1
          type: string
      additionalProperties: false
    ChatCompletionResponse:
      type: object
      properties:
        completion:
          type: string
          nullable: true
      additionalProperties: false
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          nullable: true
      additionalProperties: false
    GitHubIssueCollectionResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GitHubIssueItemResponse'
          nullable: true
      additionalProperties: false
    GitHubIssueItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
          nullable: true
        number:
          type: integer
          format: int32
          nullable: true
        title:
          type: string
          nullable: true
        body:
          type: string
          nullable: true
      additionalProperties: false
    GitHubIssueItemSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
          nullable: true
        number:
          type: integer
          format: int32
          nullable: true
        title:
          type: string
          nullable: true
        body:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
      additionalProperties: false
  securitySchemes:
    github_token:
      type: apiKey
      description: Please enter valid GitHub Token
      name: x-github-token
      in: header
    api_key:
      type: apiKey
      description: Please enter valid API Key
      name: x-webapi-key
      in: header
security:
  - github_token: []
  - api_key: []